Move all logic to get org data into the OrganizationSwitcher
